requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
azure-core==1.35.0
azure-storage-blob==12.26.0
azure-storage-file-datalake==12.21.0
blinker==1.9.0
boto3==1.40.16
botocore==1.40.16
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
comm==0.2.3
contourpy==1.3.3
cryptography==45.0.6
cycler==0.12.1
databricks-agents==1.4.0
databricks-ai-bridge==0.4.1
databricks-connect==17.1.1
databricks-langchain==0.4.2
databricks-sdk==0.64.0
databricks-vectorsearch==0.55
dataclasses-json==0.6.7
debugpy==1.8.16
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
docker==7.1.0
executing==2.2.0
fastapi==0.116.1
flask==3.1.2
fonttools==4.59.1
frozenlist==1.7.0
gitdb==4.0.12
gitpython==3.1.45
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==3.3.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
grpcio==1.74.0
grpcio-status==1.74.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
importlib-metadata==8.7.0
ipykernel==6.30.1
ipython==9.4.0
ipython-pygments-lexers==1.1.1
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyter-client==8.6.3
jupyter-core==5.8.1
kiwisolver==1.4.9
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.31
langchain-text-splitters==0.3.9
langsmith==0.4.16
mako==1.3.10
markdown==3.8.2
markdown-it-py==4.0.0
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mlflow==3.3.1
mlflow-skinny==3.3.1
mlflow-tracing==3.3.1
multidict==6.6.4
mypy-extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.3.2
openai==1.101.0
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.2
packaging==25.0
pandas==2.3.2
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt-toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
py4j==0.10.9.9
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyyaml==6.0.2
pyzmq==27.0.2
regex==2025.7.34
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rsa==4.9.1
s3transfer==0.13.1
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlalchemy==2.0.43
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.47.3
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-extensions==4.14.1
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
unitycatalog-ai==0.3.0
unitycatalog-client==0.3.0
unitycatalog-langchain==0.2.0
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
werkzeug==3.1.3
whenever==0.7.3
yarl==1.20.1
zipp==3.23.0
zstandard==0.24.0
